requirements.txt
Brotli==1.0.9
dj-database-url==0.5.0
Django==3.2.3
django-ckeditor==6.1.0
django-js-asset==1.2.2
gunicorn==20.1.0
psycopg2==2.8.4 --no-binary=psycopg2
pytz==2019.3
whitenoise==5.2.0